DL/Butterfly Species Classification/requirements.txt
keras==2.7.0
matplotlib==3.2.2
numpy==1.23.0
opencv_python==4.5.3.56
pandas==1.4.3
Pillow==9.1.1
scikit_learn==1.1.1
seaborn==0.10.1
tensorflow==2.9.1
gunicorn==20.1.0